Move RunDetails and RunExecution to their own files
