+++ src/check/runner/reporter/RunDetails.ts
@@ -1,0 +1,42 @@
+/**
+ * Post-run details produced by {@link check}
+ *
+ * A failing property can easily detected by checking the `failed` flag of this structure
+ */
+export interface RunDetails<Ts> {
+  /**
+   * Does the property failed during the execution of {@link check}?
+   */
+  failed: boolean;
+  /**
+   * Number of runs
+   *
+   * - In case of failed property: Number of runs up to the first failure (including the failure run)
+   * - Otherwise: Number of successful executions
+   */
+  numRuns: number;
+  /**
+   * Number of shrinks required to get to the minimal failing case (aka counterexample)
+   */
+  numShrinks: number;
+  /**
+   * Seed that have been used by the run
+   *
+   * It can be forced in {@link assert}, {@link check}, {@link sample} and {@link statistics} using {@link Parameters}
+   */
+  seed: number;
+  /**
+   * In case of failure: the counterexample contains the minimal failing case (first failure after shrinking)
+   */
+  counterexample: Ts | null;
+  /**
+   * In case of failure: it contains the reason of the failure
+   */
+  error: string | null;
+  /**
+   * In case of failure: path to the counterexample
+   *
+   * For replay purposes, it can be forced in {@link assert}, {@link check}, {@link sample} and {@link statistics} using {@link Parameters}
+   */
+  counterexamplePath: string | null;
+}
