+++ src/fast-check-default.ts
@@ -3,9 +3,9 @@ import 'core-js';
 import { asyncProperty } from './check/property/AsyncProperty';
 import { property } from './check/property/Property';
 import { Parameters } from './check/runner/configuration/Parameters';
+import { RunDetails } from './check/runner/reporter/RunDetails';
 import { assert, check } from './check/runner/Runner';
 import { sample, statistics } from './check/runner/Sampler';
-import { RunDetails } from './check/runner/utils/utils';
 
 import { array } from './check/arbitrary/ArrayArbitrary';
 import { boolean } from './check/arbitrary/BooleanArbitrary';
